+++ python/src/scippneutron/file_loading/_common.py
@@ -3,8 +3,10 @@
 # @author Matthew Jones
 
 from dataclasses import dataclass
-from typing import Union, Dict
+from typing import Union, Dict, Optional, Any
 import h5py
+import scipp as sc
+import numpy as np
 
 
 class BadSource(Exception):
@@ -42,3 +44,27 @@ class Group:
     parent: Union[h5py.Group, Dict]
     path: str
     contains_stream: bool = False
+
+
+def _add_attr_to_loaded_data(attr_name: str,
+                             data: sc.Variable,
+                             value: np.ndarray,
+                             unit: sc.Unit,
+                             dtype: Optional[Any] = None):
+    try:
+        data = data.attrs
+    except AttributeError:
+        pass
+
+    try:
+        if dtype is not None:
+            if dtype == sc.dtype.vector_3_float64:
+                data[attr_name] = sc.vector(value=value, unit=unit)
+            elif dtype == sc.dtype.matrix_3_float64:
+                data[attr_name] = sc.matrix(value=value, unit=unit)
+            else:
+                data[attr_name] = sc.Variable(value=value, dtype=dtype, unit=unit)
+        else:
+            data[attr_name] = sc.Variable(value=value, unit=unit)
+    except KeyError:
+        pass